+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/model/response/sync/ContactIdentity.kt
@@ -1,27 +1,21 @@
-package uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response
+package uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response.sync
 
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDateTime
 
-@Schema(description = "Phone related to a contact")
-data class ContactPhone(
-  @Schema(description = "Unique identifier for the contact phone", example = "1")
-  val contactPhoneId: Long,
+@Schema(description = "Identity related to a contact")
+data class ContactIdentity(
+  @Schema(description = "Unique identifier for the contact identity", example = "1")
+  val contactIdentityId: Long,
 
   @Schema(description = "Unique identifier for the contact", example = "123")
   val contactId: Long,
 
-  @Schema(description = "Type of phone", example = "MOBILE")
-  val phoneType: String,
+  @Schema(description = "Type of identity", example = "MOBILE")
+  val identityType: String,
 
-  @Schema(description = "Phone number", example = "+1234567890")
-  val phoneNumber: String,
-
-  @Schema(description = "Extension number", example = "123")
-  val extNumber: String?,
-
-  @Schema(description = "Indicates if this is the primary phone number", example = "true")
-  val primaryPhone: Boolean,
+  @Schema(description = "Identity ", example = "+1234567890")
+  val identityValue: String,
 
   @Schema(description = "User who created the entry", example = "admin")
   val createdBy: String,
