+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/model/response/sync/ContactPhone.kt
@@ -1,24 +1,27 @@
-package uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response
+package uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response.sync
 
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDateTime
 
-@Schema(description = "Email related to a contact")
-data class ContactEmail(
-  @Schema(description = "Unique identifier for the contact email", example = "1")
-  val contactEmailId: Long,
+@Schema(description = "Phone related to a contact")
+data class ContactPhone(
+  @Schema(description = "Unique identifier for the contact phone", example = "1")
+  val contactPhoneId: Long,
 
   @Schema(description = "Unique identifier for the contact", example = "123")
   val contactId: Long,
 
-  @Schema(description = "Type of email", example = "WORK")
-  val emailType: String,
+  @Schema(description = "Type of phone", example = "MOBILE")
+  val phoneType: String,
 
-  @Schema(description = "Email address", example = "work@example.com")
-  val emailAddress: String,
+  @Schema(description = "Phone number", example = "+1234567890")
+  val phoneNumber: String,
 
-  @Schema(description = "Indicates if this is the primary Email address", example = "true")
-  val primaryEmail: Boolean,
+  @Schema(description = "Extension number", example = "123")
+  val extNumber: String?,
+
+  @Schema(description = "Indicates if this is the primary phone number", example = "true")
+  val primaryPhone: Boolean,
 
   @Schema(description = "User who created the entry", example = "admin")
   val createdBy: String,
